+++ Editor/UTK/ImportColumn.cs
@@ -78,11 +78,18 @@ namespace UnityEditor.U2D.PSD
 
         }
 
+        string TooltipForNode(PSDTreeViewNode node)
+        {
+            return node.layer?.IsEmpty == true ? Tooltips.emptyLayerNotImportWarning : Tooltips.hiddenLayerNotImportWarning;
+        }
+
         public override void BindPSDNode(int index, PSDImporterLayerManagementMultiColumnTreeView treeView)
         {
             base.BindPSDNode(index, treeView);
             PSDTreeViewNode node = psdTreeViewNode;
             m_WarningIcon.visible = node.disable;
+            if (m_WarningIcon.visible)
+                m_WarningIcon.tooltip = TooltipForNode(node);
             if (node is PSDFoldoutTreeViewNode)
             {
                 PSDFoldoutTreeViewNode collapsable = (PSDFoldoutTreeViewNode)node;
@@ -116,7 +123,9 @@ namespace UnityEditor.U2D.PSD
             PSDTreeViewNode node = psdTreeViewNode;
             if (node != null)
             {
-                m_WarningIcon.visible = node.disable && treeView.importHidden == false;
+                m_WarningIcon.visible = node.disable && (treeView.importHidden == false || node.layer?.IsEmpty == true);
+                if (m_WarningIcon.visible)
+                    m_WarningIcon.tooltip = TooltipForNode(node);
                 if (m_ImportToggle.value != node.importLayer)
                     m_ImportToggle.SetValueWithoutNotify(node.importLayer);
                 if (node is PSDFoldoutTreeViewNode)